% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_javascript-message-handlers.R
\name{add_class}
\alias{add_class}
\title{Add CSS Class}
\usage{
add_class(.id, .class)
}
\arguments{
\item{.id}{Shiny object inputId}

\item{.class}{class to be added}
}
\value{
NA; used for side effects
}
\description{
Custom function for adding a CSS class used in {shinysurveys}.
}
\keyword{internal}
