% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-survey-code.R
\name{addPages}
\alias{addPages}
\title{Wrap Questions in the Appropriate Page Divider}
\usage{
addPages(question_df, page_id)
}
\arguments{
\item{question_df}{The data frame of questions supplied by the user,
split in \code{\link{multipaged_ui}}.}

\item{page_id}{The page ID}
}
\value{
UI for a "page" of questions
}
\description{
Wrap Questions in the Appropriate Page Divider
}
\keyword{internal}
