% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{addRequiredUI_internal}
\alias{addRequiredUI_internal}
\title{Check if a question is required}
\usage{
addRequiredUI_internal(df)
}
\arguments{
\item{df}{The output of \code{\link{nestUniqueQuestions}}.}
}
\value{
A label with or without an asterisk to signify it is required.
}
\description{
This function is for internal use. It will check if a question in the
user-supplied questions dataframe is required. If so, it will add the label
with an asterisk. If not, it will just return the label.
}
