% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-survey-code.R
\name{base_extract_user_id}
\alias{base_extract_user_id}
\title{Extract user ID from query string}
\usage{
base_extract_user_id(query_list)
}
\arguments{
\item{query_list}{The query list object from \code{shiny::parseQueryString}}
}
\value{
User ID
}
\description{
Extract user ID from query string
}
\keyword{internal}
