% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-survey-code.R
\name{base_replace_na}
\alias{base_replace_na}
\title{Replace NA values in a column using only base R}
\usage{
base_replace_na(df, column, replacement)
}
\arguments{
\item{df}{A dataframe}

\item{column}{A column name (unqouted -- similar to tidyverse syntax)}

\item{replacement}{An NA replacement value}
}
\value{
The data frame with replacements for NA in the specified column
}
\description{
Replace NA values in a column using only base R
}
\examples{

\dontrun{base_replace_na(teaching_r_questions, option, "Placeholder")}

}
\keyword{internal}
