% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-survey-code.R
\name{button_placeholders}
\alias{button_placeholders}
\title{Button placement on each page of questions}
\usage{
button_placeholders(page)
}
\arguments{
\item{page}{Current page to define UI for. Specified in \code{\link{multipaged_ui}}}
}
\value{
UI for "Next", "Previous", and "Submit" buttons
}
\description{
Button placement on each page of questions
}
\keyword{internal}
