% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{check_survey_metadata}
\alias{check_survey_metadata}
\title{Check survey metadata}
\usage{
check_survey_metadata(survey_description, survey_title)
}
\description{
Returns title/description HTML tags as needed.
}
\keyword{internal}
