% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{checkIndividual}
\alias{checkIndividual}
\title{Check if individual inputs have a value}
\usage{
checkIndividual(input = input, input_id)
}
\arguments{
\item{input}{Input from server}

\item{input_id}{The input_id to check}
}
\value{
TRUE if the input has a value; false otherwise.
}
\description{
Check if individual inputs have a value
}
