% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survey_server.R
\name{convert_type}
\alias{convert_type}
\title{Convert GUI type to proper form}
\usage{
convert_type(value)
}
\arguments{
\item{value}{The question input type}
}
\description{
Convert GUI type to proper form
}
\keyword{internal}
