% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{create_project_manual}
\alias{create_project_manual}
\title{Manual Creation of a .Rproj file}
\usage{
create_project_manual(path)
}
\arguments{
\item{path}{The file path to save .Rproj file}
}
\description{
Manual Creation of a .Rproj file
}
\keyword{internal}
