% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioMatrixInput.R
\name{create_radio_input_id}
\alias{create_radio_input_id}
\title{Create radio input ID}
\usage{
create_radio_input_id(.responseItem)
}
\arguments{
\item{.responseItem}{}
}
\value{
The response item title in a form appropriate for HTML IDs (and tidy data)
}
\description{
Create radio input ID
}
\keyword{internal}
