% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survey_server.R
\name{delete_questions}
\alias{delete_questions}
\title{Delete questions}
\usage{
delete_questions(input, question_num)
}
\arguments{
\item{input}{The Shiny input}

\item{question_num}{The question number}
}
\description{
Delete questions
}
\keyword{internal}
