% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_demo-surveys.R
\name{demo_survey}
\alias{demo_survey}
\title{Demo Survey}
\usage{
demo_survey(theme = "#63B8FF")
}
\arguments{
\item{theme}{A valid hex color such as #63B8FF (default)}
}
\value{
A Shiny App
}
\description{
This function runs a Shiny app that shows an example of running a demographic
survey in Shiny. It has a sample title and description and its theme color
can be customized using a hex color code.
}
\examples{
if (interactive()) demo_survey()

}
