% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_javascript-message-handlers.R
\name{disable_element}
\alias{disable_element}
\title{Disable HTML element}
\usage{
disable_element(.id)
}
\arguments{
\item{.id}{Shiny object inputId}
}
\value{
NA; used for side effects
}
\description{
Custom function for disabling an HTML element in {shinysurveys}.
}
\keyword{internal}
