% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survey_server.R
\name{disable_text_type_questions}
\alias{disable_text_type_questions}
\title{Get input individual question options parameter}
\usage{
disable_text_type_questions(input, question_num)
}
\arguments{
\item{input}{The Shiny input}

\item{question_num}{The question number}
}
\description{
Get input individual question options parameter
}
\keyword{internal}
