% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_question_ui.R
\name{flex_form_question_ui}
\alias{flex_form_question_ui}
\title{Create a question UI}
\usage{
flex_form_question_ui(question_number)
}
\arguments{
\item{question_number}{The question number}
}
\value{
UI
}
\description{
Create a question UI
}
