% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survey_server.R
\name{get_individual_question_options}
\alias{get_individual_question_options}
\title{Get input individual question options parameter}
\usage{
get_individual_question_options(input, question_num)
}
\arguments{
\item{input}{The Shiny input}

\item{question_num}{The question number}
}
\description{
Get input individual question options parameter
}
\keyword{internal}
