% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survey_server.R
\name{get_num_options_per_question}
\alias{get_num_options_per_question}
\title{Get number of options per question}
\usage{
get_num_options_per_question(input, question_num)
}
\arguments{
\item{input}{The Shiny input}

\item{question_num}{The question number}
}
\description{
Get number of options per question
}
\keyword{internal}
