% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survey_server.R
\name{get_question_id}
\alias{get_question_id}
\title{Get question id parameter}
\usage{
get_question_id(num_options, input, form)
}
\arguments{
\item{num_options}{A vector containing the number of options each question
has. Output from the \code{get_num_options_per_question}.}

\item{input}{The Shiny input}

\item{form}{The form reactiveValues object}
}
\description{
Get question id parameter
}
\keyword{internal}
