% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survey_server.R
\name{get_question_options}
\alias{get_question_options}
\title{Get input question options parameter}
\usage{
get_question_options(input, form)
}
\arguments{
\item{input}{The Shiny input}

\item{form}{The form reactiveValues object}
}
\description{
Get input question options parameter
}
\keyword{internal}
