% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render-survey.R
\name{getID}
\alias{getID}
\title{Get required IDs}
\usage{
getID(df)
}
\arguments{
\item{df}{The dataframe of questions}
}
\value{
The input ID for required questions
}
\description{
Get required IDs
}
\keyword{internal}
