% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{getRequired_internal}
\alias{getRequired_internal}
\title{Get a character vector of required questions}
\usage{
getRequired_internal(df)
}
\arguments{
\item{df}{One element (a dataframe) in the list of unique questions.}
}
\value{
A character vectors with the input ID of required questions.
}
\description{
Get a character vector of required questions
}
\keyword{internal}
