% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_javascript-message-handlers.R
\name{hide_survey}
\alias{hide_survey}
\title{Hide shinysurvey}
\usage{
hide_survey()
}
\value{
NA; used to hide the survey.
}
\description{
This function allows you to easily hide the survey, something you may wish to
do upon submission.
}
\examples{

if (interactive()) {

  library(shiny)
  library(shinysurveys)

  ui <- fluidPage(
    surveyOutput(teaching_r_questions,
                 survey_title = "Now you see me...",
                 survey_description = "A demo showing how to hide the survey body upon submission.")
  )

  server <- function(input, output, session) {
    renderSurvey()
    observeEvent(input$submit, hide_survey())
  }

  shinyApp(ui, server)

}

}
