% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parse-questions.R
\name{listUniqueQuestions}
\alias{listUniqueQuestions}
\title{Convert dataframe of questions to list for use in Shiny UI}
\usage{
listUniqueQuestions(df)
}
\arguments{
\item{df}{A user supplied dataframe in the format of teaching_r_questions.}
}
\value{
A list of unique questions for each UI element
}
\description{
Convert dataframe of questions to list for use in Shiny UI
}
\keyword{internal}
