% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survey_server.R
\name{make_question_dataframe}
\alias{make_question_dataframe}
\title{Make question dataframe}
\usage{
make_question_dataframe(input, form)
}
\arguments{
\item{input}{The Shiny input}

\item{form}{The form reactiveValues object}
}
\description{
Make question dataframe
}
\keyword{internal}
