% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_multipage-ui.R
\name{multipaged_ui}
\alias{multipaged_ui}
\title{Place survey questions on multiple pages}
\usage{
multipaged_ui(df)
}
\arguments{
\item{df}{The data frame of questions supplied by the user}
}
\value{
UI for all pages
}
\description{
Place survey questions on multiple pages
}
\keyword{internal}
