% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{nestUniqueQuestions}
\alias{nestUniqueQuestions}
\title{Get Unique Questions in a Nested Dataframe}
\usage{
nestUniqueQuestions(df)
}
\arguments{
\item{df}{A user supplied dataframe in the format of teaching_r_questions.}
}
\value{
A nested dataframe.
}
\description{
Get Unique Questions in a Nested Dataframe
}
