% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-survey-code.R
\name{parse_num_ranks}
\alias{parse_num_ranks}
\title{Parse the number of ranks to include for rank questions}
\usage{
parse_num_ranks(input_type)
}
\arguments{
\item{input_type}{The input type}
}
\value{
Numeric to be used for determining number of ranks
}
\description{
Parse the number of ranks to include for rank questions
}
\keyword{internal}
