% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parse-questions.R
\name{pluck_by_index}
\alias{pluck_by_index}
\title{Simple pluck}
\usage{
pluck_by_index(list, index)
}
\arguments{
\item{list}{A list object}

\item{index}{A numeric index}
}
\value{
Returns list element.
}
\description{
simple and specific version of purrr::pluck to meet use-case
}
\keyword{internal}
