% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_radioMatrixInput.R
\name{radioBody}
\alias{radioBody}
\title{Create the table body}
\usage{
radioBody(.responseItems, .choices, .selected = NULL)
}
\arguments{
\item{.responseItems}{Questions to be asked (row labels)}

\item{.choices}{Possible choices (values for radio buttons)}

\item{.selected}{Initial selected value}
}
\value{
UI for the matrix input (table) body
}
\description{
Create the table body
}
\keyword{internal}
