% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioMatrixInput.R
\name{radioMatHeader}
\alias{radioMatHeader}
\title{Create the radio matrix input's header}
\usage{
radioMatHeader(.choices)
}
\arguments{
\item{.choices}{Possible choices}
}
\value{
Header for the table (matrix input)
}
\description{
Create the radio matrix input's header
}
\keyword{internal}
