% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioMatrixInput.R
\name{radioMatrixButtons}
\alias{radioMatrixButtons}
\title{Create the actual radio button inputs}
\usage{
radioMatrixButtons(inputId, choices, selected = NULL)
}
\arguments{
\item{inputId}{This is the ID for the question to which the radio button
inputs correspond}

\item{choices}{The choices (values) for each radio button to indicate}

\item{selected}{A default selected value}
}
\value{
radio button input UI
}
\description{
Create the actual radio button inputs
}
\keyword{internal}
