% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-survey-code.R
\name{rank_ui_internal}
\alias{rank_ui_internal}
\title{Create rank UI for individual options}
\usage{
rank_ui_internal(df, num_ranks)
}
\arguments{
\item{df}{Individual option rows of the df}

\item{num_ranks}{The number of ranking}
}
\value{
UI for individual rank elements
}
\description{
Create rank UI for individual options
}
\keyword{internal}
