% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_javascript-message-handlers.R
\name{remove_class}
\alias{remove_class}
\title{Remove CSS Class}
\usage{
remove_class(.id, .class)
}
\arguments{
\item{.id}{Shiny object inputId}

\item{.class}{class to be removed}
}
\value{
NA; used for side effects
}
\description{
Custom function for removing a CSS class used in {shinysurveys}.
}
\keyword{internal}
