% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survey_server.R
\name{remove_questions}
\alias{remove_questions}
\title{Remove questions}
\usage{
remove_questions(input, question_num)
}
\arguments{
\item{input}{The Shiny input}

\item{question_num}{The question number}
}
\description{
Remove questions
}
\keyword{internal}
