% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{renderSurvey}
\alias{renderSurvey}
\title{Server code for adding survey questions}
\usage{
renderSurvey(input, output, df, session, theme = "#63B8FF")
}
\arguments{
\item{input}{Input from server}

\item{output}{Output from server}

\item{df}{A user supplied dataframe in the format of teaching_r_questions.}

\item{session}{Session from server}

\item{theme}{A valid hex color such as #63B8FF (default)}
}
\value{
NA; server code
}
\description{
Create the UI code for a Shiny app based on user-supplied questions. Possible
question (input) types include numeric, text, multiple choice, or selection.
}
