% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render-survey.R
\name{showDependence}
\alias{showDependence}
\title{Show dependence questions}
\usage{
showDependence(input = input, df)
}
\arguments{
\item{input}{Input from server}

\item{df}{One element (a dataframe) in the list of unique questions.}
}
\value{
NA; shows a dependence question in the UI.
}
\description{
Show dependence questions
}
\keyword{internal}
