% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_shinysurveys.R
\name{surveyID}
\alias{surveyID}
\title{Add correct ID for custom input types}
\usage{
surveyID()
}
\value{
NA; used for side effects with \code{\link{extendInputType}}.
}
\description{
\code{surveyID()} is a helper function for \code{\link{extendInputType}}. When
defining custom input types, the \code{inputId} argument for shiny UI components
should equal \code{surveyID()}. See examples for more details.
}
\examples{

extendInputType("slider", {
shiny::sliderInput(
  inputId = surveyID(),
  label = surveyLabel(),
  min = 1,
  max = 10,
  value = 5
)
})

}
\seealso{
\code{\link{extendInputType}}

\code{\link{surveyLabel}}

\code{\link{surveyOptions}}
}
