% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_extend-shinysurveys.R
\name{surveyLabel}
\alias{surveyLabel}
\title{Add correct label for custom input types}
\usage{
surveyLabel()
}
\value{
NA; used for side effects with \code{\link{extendInputType}}.
}
\description{
\code{surveyLabel()} is a helper function for \code{\link{extendInputType}}. When
defining custom input types, the \code{label} argument for shiny UI components
should equal \code{surveyLabel()}. It essentially takes on the value in the
"question" column in the data supplied to \code{\link{surveyOutput}}. See
examples for more details.
}
\examples{

extendInputType("slider", {
shiny::sliderInput(
  inputId = surveyID(),
  label = surveyLabel(),
  min = 1,
  max = 10,
  value = 5
)
})

}
\seealso{
\code{\link{extendInputType}}

\code{\link{surveyID}}

\code{\link{surveyOptions}}
}
