% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_shinysurveys.R
\name{surveyOptions}
\alias{surveyOptions}
\title{Add options for custom input types}
\usage{
surveyOptions()
}
\value{
NA; used for side effects with \code{\link{extendInputType}}.
}
\description{
\code{surveyOptions()} is a helper function for \code{\link{extendInputType}}. When
defining custom input types, the choices arguments for shiny UI components
should equal \code{surveyOption()}. See examples for more details.
}
\examples{

extendInputType("inlineRadioButtons", {
shiny::radioButtons(
  inputId = surveyID(),
  label = surveyLabel(),
  selected = character(0),
  choices = surveyOptions(),
  inline = TRUE
)
})

}
\seealso{
\code{\link{extendInputType}}

\code{\link{surveyID}}

\code{\link{surveyOptions}}
}
