% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survey-output.R
\name{surveyOutput_individual}
\alias{surveyOutput_individual}
\title{Generate the UI Code for demographic questions}
\usage{
surveyOutput_individual(df)
}
\arguments{
\item{df}{One element (a dataframe) in the list of unique questions.}
}
\value{
UI Code for a Shiny App.
}
\description{
Generate the UI Code for demographic questions
}
\keyword{internal}
