% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teaching_r_questions}
\alias{teaching_r_questions}
\title{A sample CSV file for demographic questions}
\format{
A data frame with 54 rows and 6 columns:
\describe{
\itemize{
\item question: The question to be asked.
\item option: A possible response to the question.
In multiple choice questions, for example, this would be the possible answers.
For questions without discrete answers, this would be the default option shown on the input.
\item input_type: What type of response is expected? Numeric, multiple choice, text, etc...
\item input_id: The input id for Shiny inputs.
\item dependence: Does this question (row) depend on another?
That is, should it only appear if a different question has a specific value?
This column contains the input_id of whatever question this one depends upon.
\item dependence_value: This column contains the specific value that the dependence
question must take for this question (row) to be shown.
\item required: logical TRUE/FALSE signifying if a question is required.
}
}
}
\source{
D'Agostino McGowan Data Science Lab at Wake Forest University.
}
\usage{
teaching_r_questions
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
