% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey-output.R
\name{title_placeholder}
\alias{title_placeholder}
\title{Control Title UI Placement for Multi-paged Surveys}
\usage{
title_placeholder(page)
}
\arguments{
\item{page}{}
}
\value{
UI for title if applicable
}
\description{
Control Title UI Placement for Multi-paged Surveys
}
\keyword{internal}
