% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render-survey.R
\name{toggle_element}
\alias{toggle_element}
\title{Toggle element state}
\usage{
toggle_element(id, condition)
}
\arguments{
\item{id}{Shiny object inputId}

\item{condition}{Condition on which to enable or disable}
}
\value{
NA; used for side effects
}
\description{
Custom function for toggling enable/disable state of HTML element in {shinysurveys}.
}
\keyword{internal}
