% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_survey.R
\name{demo_survey}
\alias{demo_survey}
\title{Demo Survey over multiple pages}
\usage{
demo_survey()
}
\value{
A Shiny App
}
\description{
This function runs a Shiny app that shows an example of running a demographic
survey in Shiny. Upon submission, it saves the responses to your downloads
folder.
}
\details{
It requires shiny, shinyjs, and shinyWidgets.
}
\examples{

\dontrun{
demo_survey()
}
}
