% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-survey-code.R
\name{get_questions}
\alias{get_questions}
\title{Get unique questions from user-input dataframe}
\usage{
get_questions(df)
}
\arguments{
\item{df}{User-input dataframe of questions}
}
\value{
List of questions
}
\description{
Get unique questions from user-input dataframe
}
\keyword{internal}
