% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{getSurveyData}
\alias{getSurveyData}
\title{Get survey data}
\usage{
getSurveyData(custom_id = NULL)
}
\arguments{
\item{custom_id}{A unique identifier for the survey's respondents. NULL by
default, and the built-in {shinysurveys} userID will be used.}
}
\value{
A participant's responses.
}
\description{
Get a participant's responses.
}
\examples{

if (interactive()) {

 library(shiny)

 ui <- fluidPage(
   surveyOutput(teaching_r_questions)
 )

 server <- function(input, output, session) {
   renderSurvey()
   # Upon submission, print a data frame with participant responses
   observeEvent(input$submit, {
     print(getSurveyData())
   })
 }

 shinyApp(ui, server)

 }

}
