% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{getRequired_internal}
\alias{getRequired_internal}
\title{Get a character vector of required questions}
\usage{
getRequired_internal(questions)
}
\arguments{
\item{questions}{The list of unique questions from \code{\link{listUniqueQuestions}}.}
}
\value{
A character vectors with the input ID of required questions.
}
\description{
Get a character vector of required questions
}
\keyword{internal}
