% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_radioMatrixInput.R
\name{radioMatHeader}
\alias{radioMatHeader}
\title{Create the radio matrix input's header}
\usage{
radioMatHeader(.choices, .required)
}
\arguments{
\item{.choices}{Possible choices}

\item{.required}{Logical: TRUE/FALSE should a required asterisk be placed on the matrix question}
}
\value{
Header for the table (matrix input)
}
\description{
Create the radio matrix input's header
}
\keyword{internal}
