% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{renderSurvey}
\alias{renderSurvey}
\title{Server code for adding survey questions}
\usage{
renderSurvey(df, theme = "#63B8FF")
}
\arguments{
\item{df}{A user supplied dataframe in the format of teaching_r_questions.}

\item{theme}{A valid R color: predefined such as "red" or "blue"; hex colors such as #63B8FF (default)}
}
\value{
NA; used for server-side logic in Shiny apps.
}
\description{
Include server-side logic for shinysurveys.
}
\examples{

if (interactive()) {
renderSurvey(df = shinysurveys::teaching_r_questions,
theme = "#63B8FF")
}

}
