% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{renderSurvey}
\alias{renderSurvey}
\title{Server code for adding survey questions}
\usage{
renderSurvey(df, input, output, session, theme = "#63B8FF")
}
\arguments{
\item{df}{A user supplied dataframe in the format of teaching_r_questions.}

\item{input}{Input from server}

\item{output}{Output from server}

\item{session}{Session from server}

\item{theme}{A valid hex color such as #63B8FF (default)}
}
\value{
NA; server code
}
\description{
Include server-side logic for shinysurveys.
}
\examples{

\dontrun{
renderSurvey(df = shinysurveys::teaching_r_questions,
input = input,
output = output,
session = session,
theme = "#63B8FF")
}

}
