% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{showDependence}
\alias{showDependence}
\title{Show dependence questions}
\usage{
showDependence(input = input, df)
}
\arguments{
\item{input}{Input from server}

\item{df}{The output of \code{\link{nestUniqueQuestions}} (indexed into the data column)..}
}
\value{
NA; shows a dependence question in the UI.
}
\description{
Show dependence questions
}
