% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parse-questions.R
\name{split_dependence}
\alias{split_dependence}
\title{Split questions based on dependency}
\usage{
split_dependence(df)
}
\arguments{
\item{df}{A data frame subset for one question}
}
\value{
A list where each element is one UI question.
}
\description{
Split questions based on dependency
}
\keyword{internal}
