% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{surveyOutput}
\alias{surveyOutput}
\title{Generate the UI Code for demoraphic questions}
\usage{
surveyOutput(df, ...)
}
\arguments{
\item{df}{A user supplied dataframe in the format of teaching_r_questions.}

\item{...}{Additional arguments to pass into \link[shiny]{actionButton} used to submit survey responses.}
}
\value{
UI Code for a Shiny App.
}
\description{
Create the UI code for a Shiny app based on user-supplied questions. Possible
question (input) types include numeric, text, multiple choice, or selection.
}
\examples{
\dontrun{
surveyOutput(teaching_r_questions)
}
}
