% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{surveyOutput_individual}
\alias{surveyOutput_individual}
\title{Generate the UI Code for demoraphic questions}
\usage{
surveyOutput_individual(df)
}
\arguments{
\item{df}{A nested dataframe.}
}
\value{
UI Code for a Shiny App.
}
\description{
Generate the UI Code for demoraphic questions
}
\keyword{internal}
