% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_code.R
\name{surveyOutput_individual}
\alias{surveyOutput_individual}
\title{Generate the UI Code for demographic questions}
\usage{
surveyOutput_individual(df, custom_input, ...)
}
\arguments{
\item{df}{One element (a dataframe) in the list of unique questions.}

\item{custom_input}{The input_type corresponding to the custom input.}

\item{...}{Additional arguments to pass into input_extension()}
}
\value{
UI Code for a Shiny App.
}
\description{
Generate the UI Code for demographic questions
}
\keyword{internal}
